Moved the test-framework patch to a separate file
Only patch the version 2.x.x of Jasmine
