+++ lib/Unexpected.js
@@ -1,4 +1,4 @@
-/*global Promise:true, it:true, jasmine*/
+/*global Promise:true*/
 var Assertion = require('./Assertion');
 var createStandardErrorMessage = require('./createStandardErrorMessage');
 var utils = require('./utils');
@@ -11,6 +11,7 @@ var makePromise = require('./makePromise');
 var oathbreaker = require('./oathbreaker');
 var throwIfNonUnexpectedError = require('./throwIfNonUnexpectedError');
 var UnexpectedError = require('./UnexpectedError');
+var testFrameworkPatch =require('./testFrameworkPatch');
 
 var anyType = {
     name: 'any',
@@ -41,81 +42,9 @@ var anyType = {
     }
 };
 
-function jasmineFail(err) {
-    if (typeof jasmine === 'object') {
-        jasmine.getEnv().fail(err);
-    }
-}
-
-function jasmineSuccess(err) {
-    if (typeof jasmine === 'object') {
-        jasmine.getEnv().expect(true).toBe(true);
-    }
-}
-
-var promiseCreated = false;
-
-function patchItMethod() {
-    if (typeof it === 'function' && !it.isPatched) {
-        var originalIt = it;
-        it = function (title, fn) {
-            if (!fn) {
-                return originalIt(title);
-            }
-            var async = fn.length > 0;
-            var wrapper = function (done) {
-                promiseCreated = false;
-                var result;
-                try {
-                    if (async) {
-                        fn.call(this, function (err) {
-                            if (err) {
-                                jasmineFail(err);
-                                done(err);
-                            } else {
-                                jasmineSuccess();
-                                done();
-                            }
-                        });
-                        return;
-                    } else {
-                        result = fn.call(this);
-                    }
-                    var isPromise = result && typeof result === 'object' && typeof result.then === 'function';
-                    if (isPromise) {
-                        result.then(function () {
-                            jasmineSuccess();
-                            done();
-                        }).caught(function (err) {
-                            jasmineFail(err);
-                            done(err);
-                        });
-                    } else if (promiseCreated) {
-                        throw new Error('When using asynchronous assertions you must return a promise from the it block');
-                    } else {
-                        jasmineSuccess();
-                        done();
-                    }
-                } catch (err) {
-                    jasmineFail(err);
-                    return done(err);
-                }
-            };
-            wrapper.toString = function () {
-                return fn.toString();
-            };
-            return originalIt(title, wrapper);
-        };
-
-        Object.keys(originalIt).forEach(function (methodName) {
-            it[methodName] = originalIt[methodName];
-        });
-        it.isPatched = true;
-    }
-}
 
 function Unexpected(options) {
-    patchItMethod();
+    testFrameworkPatch.applyPatch();
     options = options || {};
     this.assertions = options.assertions || {any: {}};
     this.typeByName = options.typeByName || {};
@@ -651,7 +580,7 @@ Unexpected.prototype.expect = function expect(subject, testDescriptionString) {
                 try {
                     var result = oathbreaker(callback());
                     if (result && typeof result.then === 'function' && typeof result.caught === 'function') {
-                        promiseCreated = true;
+                        testFrameworkPatch.promiseCreated();
                         return result.caught(function (e) {
                             if (e && e._isUnexpected) {
                                 truncateStack(e, wrappedExpect);
@@ -790,7 +719,7 @@ Unexpected.prototype.expect = function expect(subject, testDescriptionString) {
     try {
         var promise = executeExpect(subject, testDescriptionString, args);
         if (promise && typeof promise.then === 'function') {
-            promiseCreated = true;
+            testFrameworkPatch.promiseCreated();
             return promise.caught(function (e) {
                 if (e && e._isUnexpected) {
                     // TODO truncate the stack
